Removed channel endpoints
